+++ test/Unit/Ads/OpenMeasurement/OpenMeasurementTest.ts
@@ -195,7 +195,6 @@ import { ThirdPartyEventManager } from 'Ads/Managers/ThirdPartyEventManager';
 
                             sinon.stub(deviceInfo, 'getScreenWidth').resolves(1280);
                             sinon.stub(deviceInfo, 'getScreenHeight').resolves(768);
-                            sandbox.stub(om, 'loaded');
                             sandbox.stub(om, 'geometryChange');
                             sandbox.stub(om, 'impression');
 
@@ -204,7 +203,6 @@ import { ThirdPartyEventManager } from 'Ads/Managers/ThirdPartyEventManager';
 
                         it('should call session start ad events', () => {
                             return om.onEventProcessed('sessionStart').then(() => {
-                                sinon.assert.called(<sinon.SinonSpy>om.loaded);
                                 sinon.assert.called(<sinon.SinonSpy>om.impression);
                                 sinon.assert.notCalled(<sinon.SinonSpy>om.geometryChange);
                             });
@@ -218,7 +216,6 @@ import { ThirdPartyEventManager } from 'Ads/Managers/ThirdPartyEventManager';
                                     clock.tick(1000);
                                     clock.restore();
                                     sinon.assert.called(<sinon.SinonSpy>om.impression);
-                                    sinon.assert.called(<sinon.SinonSpy>om.loaded);
                                     sinon.assert.called(<sinon.SinonSpy>om.geometryChange);
                                 });
                             });